Add CognitiveCoreService and update tests
