+++ src/deepthought/modules/input_handler.py
@@ -1,4 +1,3 @@
-# File: src/deepthought/modules/input_handler.py
 import logging
 import uuid
 from datetime import datetime, timezone
@@ -7,7 +6,6 @@ import nats
 from nats.aio.client import Client as NATS
 from nats.js.client import JetStreamContext
 
-# Assuming eda modules are in parent dir relative to modules dir
 from ..eda.events import EventSubjects, InputReceivedPayload, MemoryRetrievedPayload
 from ..eda.publisher import Publisher
 
@@ -17,10 +15,10 @@ logger = logging.getLogger(__name__)
 class InputHandler:
     """Handles user input and publishes InputReceived events."""
 
-    def __init__(self, nats_client: NATS, js_context: JetStreamContext, hierarchical_service=None):
-        """Initialize with optional hierarchical memory service."""
+    def __init__(self, nats_client: NATS, js_context: JetStreamContext, cognitive_service=None):
+        """Initialize with optional cognitive memory service."""
         self._publisher = Publisher(nats_client, js_context)
-        self._memory_service = hierarchical_service
+        self._memory_service = cognitive_service
         logger.info("InputHandler initialized (JetStream enabled).")
 
     async def process_input(self, user_input: str) -> str:
@@ -29,11 +27,9 @@ class InputHandler:
             raise ValueError("user_input must be a string")
 
         input_id = str(uuid.uuid4())
-        # Use timezone-aware UTC timestamp
         timestamp = datetime.now(timezone.utc).isoformat()
         payload = InputReceivedPayload(user_input=user_input, input_id=input_id, timestamp=timestamp)
         try:
-            # Always use JetStream for input events in this version
             await self._publisher.publish(
                 EventSubjects.INPUT_RECEIVED,
                 payload,
