+++ src/deepthought/services/__init__.py
@@ -1,19 +1,13 @@
 """Service utilities for DeepThought."""
 
+from .cognitive_core_service import CognitiveCoreService
 from .db_manager import DBManager
 from .file_graph_dal import FileGraphDAL
-from .hierarchical_service import HierarchicalService
-from .knowledge_graph_service import KnowledgeGraphService
-from .memory_service import MemoryService
 from .persona_manager import PersonaManager
-from .social_graph_service import SocialGraphService
 
 __all__ = [
     "FileGraphDAL",
-    "MemoryService",
-    "HierarchicalService",
-    "KnowledgeGraphService",
+    "CognitiveCoreService",
     "PersonaManager",
-    "SocialGraphService",
     "DBManager",
 ]
