+++ tests/unit/modules/test_input_handler.py
@@ -1,8 +1,12 @@
 import json
+import sys
+import types
 from datetime import datetime, timezone
 from types import SimpleNamespace
 
 import pytest
+
+sys.modules.setdefault("aiohttp", types.ModuleType("aiohttp"))
 
 pytest.importorskip("nats")
 
@@ -48,7 +52,7 @@ async def test_process_input_success():
     js = DummyJS()
     nc = DummyNATS()
     memory = DummyMemory()
-    handler = InputHandler(nc, js, hierarchical_service=memory)
+    handler = InputHandler(nc, js, cognitive_service=memory)
     input_id = await handler.process_input("hello")
 
     assert js.published
